refactor offchain adapter and cmd
