+++ src/interfaces/cli/commands/offchain-cmd.js
@@ -1,9 +1,12 @@
 const inquirer = require("inquirer");
+const { sha3 } = require("tribute-contracts/utils/ContractUtil");
+const { configs } = require("../../../../cli-config");
 
 const {
   submitOffchainResult,
   newOffchainVote,
 } = require("../../../contracts/adapters/offchain-voting-adapter");
+const { logEnvConfigs, success, notice, info } = require("../../../utils/logging");
 
 const offchainCommands = (program) => {
   program
@@ -21,9 +24,28 @@ const offchainCommands = (program) => {
             choices: ["Yes", "No"],
           },
         ])
-        .then((vote) =>
-          newOffchainVote(snapshotProposalId, vote.choice, data, program.opts())
-        );
+        .then((vote) => {
+          const daoProposalId = sha3(snapshotProposalId);
+
+          notice(`\n ::: Submitting offchain voting...\n`);
+          logEnvConfigs(configs, configs.contracts.OffchainVoting);
+          info(`Snapshot Proposal Id:\t${snapshotProposalId}`);
+          info(`DAO Proposal Id:\t${daoProposalId}`);
+          info(`Choice:\t\t\t${vote.choice}`);
+          info(`Data:\t\t\t${data ? data : "n/a"}\n`);
+
+          return newOffchainVote(
+            snapshotProposalId,
+            daoProposalId,
+            vote.choice,
+            data
+          );
+        })
+        .then((res) => {
+          notice(`Member ${res.memberAddress} has voted "${res.choice}"\n`);
+          notice(`Snapshot Proposal Id ${snapshotProposalId}\n`);
+          success("::: Offchain vote submitted to Snapshot Hub.");
+        });
     });
 
   program
@@ -31,7 +53,22 @@ const offchainCommands = (program) => {
     .description(
       "Submit an offchain vote result to the DAO using the snapshot proposal id."
     )
-    .action((snapshotProposalId) => submitOffchainResult(snapshotProposalId));
+    .action((snapshotProposalId) => {
+      const daoProposalId = sha3(snapshotProposalId);
+
+      notice(`\n ::: Submitting offchain voting results...\n`);
+      logEnvConfigs(configs, configs.contracts.OffchainVotingContract);
+
+      info(`Snapshot Proposal Id:\t${snapshotProposalId}`);
+      info(`DAO Proposal Id:\t${daoProposalId}`);
+
+      return submitOffchainResult(snapshotProposalId, daoProposalId).then(
+        (res) => {
+          notice(`DAO Proposal Id ${res.daoProposalId}\n`);
+          success(`::: Offchain vote results submitted to the DAO`);
+        }
+      );
+    });
 
   return program;
 };
